% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect.R
\name{autoselect_profile}
\alias{autoselect_profile}
\title{Automatically Select a Profile}
\usage{
autoselect_profile(root_dir, name = "default")
}
\arguments{
\item{root_dir}{The root directory where Firefox' data is stored. See
'Details' in the documentation of \code{\link[=connect_history_db]{connect_history_db()}}
for hints where to find this on different systems.}

\item{name}{character giving the name of the profile to be selected.}
}
\value{
a character vector of length 1 giving the path to the selected profile.
}
\description{
Automatically select one of the available profiles. If a profile with the
name given by the argument \code{name} exists, its path is returned.
Otherwise, the path of the first available profile is returned.
}
\seealso{
List all available profiles with \code{\link[=list_profiles]{list_profiles()}}.
}
