% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/browserhistory-package.R
\docType{package}
\name{browserhistory-package}
\alias{browserhistory}
\alias{browserhistory-package}
\title{Read and Analyse Firefox' Browser History}
\description{
This package provides functionality to read and analyse the browser history
of Firefox. The history is stored in a SQLite database and can be accessed
using \code{\link[=read_browser_history]{read_browser_history()}}.
}
\details{
Other browsers might be supported in the future.
}
\author{
\strong{Maintainer}: Stefan Lanz \email{slanz1137@gmail.com}

}
\keyword{internal}
