% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect.R
\name{connect_history_db}
\alias{connect_history_db}
\title{Connect to History Database}
\usage{
connect_history_db(root_dir, profile)
}
\arguments{
\item{root_dir}{The root directory where Firefox' data is stored. See
'Details' for hints where to find this on different systems.}

\item{profile}{The profile name to be used. The profile is a folder
within the root dir. The names of the available profiles can be found
in the file \code{profiles.ini}.}
}
\description{
Create a Connection to Firefox' history database.
}
\details{
Where Firefox stores the history database depends on your opperating system.
This has been checked only for very few systems:

\describe{
\item{Ubuntu 22.04}{If Firefox has been installed from a deb package, the
root directory is \verb{~/.mozilla/firefox}. (This might also be true for other
Debian-based distributions.) If Firefox has been installed as a snap
package, the root directory is \verb{~/snap/firefox/common/.mozilla/firefox}}
}
}
