% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_data.R
\name{generate_testdata}
\alias{generate_testdata}
\title{Generate Test Data}
\usage{
generate_testdata(
  con = NULL,
  root_dir = NULL,
  profile = autoselect_profile(root_dir),
  start_time,
  output_dir
)
}
\arguments{
\item{con}{a connection to a history database created with
\code{\link[=connect_history_db]{connect_history_db()}}.}

\item{root_dir}{The root directory where Firefox' data is stored. See
'Details' in the documentation of \code{\link[=connect_history_db]{connect_history_db()}}
for hints where to find this on different systems.}

\item{profile}{The profile name to be used. The profile is a folder
within the root directory. The names of the available profiles can be listed
using \code{\link[=list_profiles]{list_profiles()}}. If no profile is specified, the function
automatically selects a profile called "default", if it is available,
and the first available profile otherwise.}

\item{start_time}{a POSIXt object that specifies the start time of the data.
All elements of the history that are newer than this date will be included
in the test data.}

\item{output_dir}{The directory where the test database should be
written to. The directory must exist.}
}
\value{
The function reads the test database it has written and returns
its contents as a data frame.
}
\description{
Generate a test database from a Firefox history database.
}
\details{
A file \code{profiles.ini} will be created in the output directory that contains
the test profile. A folder \code{test_profile} will be created that contains
the test database in a file called \code{places.sqlite}. The database contains
only the three tables (\code{moz_historyvisits}, \code{moz_places}, \code{moz_origins})
that are processed by \code{\link[=read_browser_history]{read_browser_history()}}.
}
