% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect.R
\name{guess_root_dir}
\alias{guess_root_dir}
\title{Try to Guess the Directory Where the Firefox History Is Stored}
\usage{
guess_root_dir()
}
\value{
a character vector with the path to the root directory or \code{NA} if
guessing was unsuccessful.
}
\description{
Try to guess the directory, where Firefox stores the history database. This
works according to the rules laid out in the documentation of
\code{\link[=connect_history_db]{connect_history_db()}}. If it returns a directory path, this path is
guaranteed to exist. If guessing is unsuccessful, \code{NA} is returned.
}
