% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_browser_history}
\alias{read_browser_history}
\title{Read History From Database}
\usage{
read_browser_history(
  con = NULL,
  root_dir = guess_root_dir(),
  profile = autoselect_profile(root_dir),
  start_time = NULL,
  end_time = NULL,
  tz = "",
  raw = FALSE
)
}
\arguments{
\item{con}{a connection to a history database created with
\code{\link[=connect_history_db]{connect_history_db()}}. Instead of passing a connection, you can also
specify a root directory and (optionally) a profile and the function
will build up the connection automatically.}

\item{root_dir}{The root directory where Firefox' data is stored. See
'Details' in the documentation of \code{\link[=connect_history_db]{connect_history_db()}}
for hints where to find this on different systems. If omitted,
the function tries to guess the root directory using \code{\link[=guess_root_dir]{guess_root_dir()}}.}

\item{profile}{The profile name to be used. The profile is a folder
within the root directory. The names of the available profiles can be listed
using \code{\link[=list_profiles]{list_profiles()}}. If no profile is specified, the function
automatically selects a profile called "default", if it is available,
and the first available profile otherwise.}

\item{start_time, end_time}{POSIXt objects that specifies the time interval
for which browser history will be returned.}

\item{tz}{character giving the time zone to be used for the conversion of
the timestamps. The default value \code{""} is the current time zone.
See \link[base:timezones]{time zones} for more information on time zones in R.}

\item{raw}{logical. Should a raw table with all columns be returned?}
}
\value{
a \code{tibble} with the browser history
}
\description{
Read the Browser History from the database. The function can be called
with a database connection created with \code{\link[=connect_history_db]{connect_history_db()}} or by
specifying a root directory and (optionally).
}
