% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect.R
\name{autoselect_profile}
\alias{autoselect_profile}
\title{Automatically Select a Profile}
\usage{
autoselect_profile(root_dir)
}
\arguments{
\item{root_dir}{The root directory where Firefox' data is stored. See
'Details' in the documentation of \code{\link[=connect_history_db]{connect_history_db()}}
for hints where to find this on different systems.}
}
\value{
a character vector of length 1 giving the path to the selected profile.
}
\description{
Automatically select one of the available profiles. If a profile is marked
as default (\code{Default=1}), its path is returned.
Otherwise, the path of the profile labelled \code{Profile0} is returned. If that
also does not exist, the first profile is returned. If no profiles are
defined, the function throws an error.
}
\seealso{
List all available profiles with \code{\link[=list_profiles]{list_profiles()}}.
}
