% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/browserhistory-package.R
\docType{package}
\name{browserhistory-package}
\alias{browserhistory}
\alias{browserhistory-package}
\title{browserhistory: Read and Analyse Firefox' Browser History}
\description{
Tools to read and analyse Firefox' browser history. Support for other browsers may be added in the future.
}
\author{
\strong{Maintainer}: Stefan Lanz \email{slanz1137@gmail.com}

}
\keyword{internal}
