% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect.R
\name{connect_history_db}
\alias{connect_history_db}
\title{Connect to History Database}
\usage{
connect_history_db(root_dir, profile = autoselect_profile(root_dir))
}
\arguments{
\item{root_dir}{The root directory where Firefox' data is stored. See
'Details' for hints where to find this on different systems.}

\item{profile}{The profile name to be used. The profile is a folder
within the root directory. The names of the available profiles can be listed
using \code{\link[=list_profiles]{list_profiles()}}. If no profile is specified, the function
automatically selects a profile called "default", if it is available,
and the first available profile otherwise.}
}
\value{
a database connection (class \code{SQLiteConnection}) to the history database
}
\description{
Create a Connection to Firefox' history database.
}
\details{
Where Firefox stores the history database depends on your operating system.
This has been checked only for very few systems:

\describe{
\item{Ubuntu 22.04}{If Firefox has been installed from a deb package, the
root directory is \verb{~/.mozilla/firefox}. (This might also be true for other
Debian-based distributions.) If Firefox has been installed as a snap
package, the root directory is \verb{~/snap/firefox/common/.mozilla/firefox}.}
}
}
